separating out the helpers
