+++ provisioning/service/src/Microsoft.Azure.Devices.Provisioning.Service.csproj
@@ -50,7 +50,7 @@
     <Compile Include="$(common)\service\StringValidationHelper.cs" Link="Common\StringValidationHelper.cs" />
     <Compile Include="$(common)\service\Utils.cs" Link="Common\Utils.cs" />
     <Compile Include="$(Common)\DefaultWebProxySettings.cs" />
-    <Compile Include="$(Common)\HttpContentHelper.cs" />
+    <Compile Include="$(Common)\HttpContentExtensions.cs" />
   </ItemGroup>
 
   <ItemGroup>
