+++ provisioning/transport/http/src/Microsoft.Azure.Devices.Provisioning.Transport.Http.csproj
@@ -83,7 +83,7 @@
     <Compile Include="$(Generated)\Models\DeviceRegistration.cs">
       <Link>Generated\Models\DeviceRegistration.cs</Link>
     </Compile>
-    <Compile Include="$(Common)\HttpContentHelper.cs">
+    <Compile Include="$(Common)\HttpContentExtensions.cs">
       <LinkBase>Common</LinkBase>
     </Compile>
   </ItemGroup>
